% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConcepts.R
\docType{package}
\name{CreateConceptDatasets}
\alias{CreateConceptDatasets}
\title{This is my first package}
\usage{
CreateConceptDatasets(
  codesheet,
  file,
  c.voc,
  c.concept,
  c.codes,
  c.startwith,
  f.code,
  f.voc,
  path,
  method = "SQL"
)
}
\arguments{
\item{codesheet}{data.table file with the codes, concept and coding system}

\item{file}{CDM table with codes}

\item{c.voc}{codesheet coding system column name}

\item{c.concept}{codesheet concept column name}

\item{c.codes}{codesheet code column name}

\item{f.code}{file code column name}

\item{f.voc}{file coding system column name}

\item{path}{location to write the RDS files to per concept name}

\item{method}{if SQL, a join is used. If loop a loop is used to subset}

\item{coding}{systems that merges based on start with}
}
\value{
RDS files to per concept name with the rows that have the code specified in the codesheet
}
\section{??}{
NA
}

\author{
Roel Elbers
}
\keyword{??}
