% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLargeSimulatedDataset.R
\docType{package}
\name{CreateLargeSimulatedDataset}
\alias{CreateLargeSimulatedDataset}
\title{This is my second function I add to the package package}
\usage{
CreateLargeSimulatedDataset(
  Inputfolder,
  Outputfolder,
  N,
  Identifier_name,
  Delimiter
)
}
\arguments{
\item{Inputfolder}{path where the csv files are located}

\item{Outputfolder}{path where the enlarged csv files need to be stored}

\item{N}{factor of enlargement}

\item{Identifier_name}{Column name(s) that needs to be extended to ensure unique id's}

\item{Delimiter}{Define a separator for csv files}
}
\value{
Csv files N times the number of rows as the csv files in the input folder
}
\section{??}{
NA
}

\author{
Roel Elbers
}
\keyword{??}
